@@ -22,68 +22,68 @@
     "format:write": "nx format:write",
     "format:check": "nx format:check",
     "update": "nx migrate latest",
-    "workspace-schematic": "nx workspace-schematic",
+    "workspace-generator": "nx workspace-generator",
     "dep-graph": "nx dep-graph",
     "help": "nx help"
   },
   "private": true,
   "dependencies": {
-    "rxjs": "^6.5.4",
-    "tslib": "^1.10.0"
-  },
-  "peerDependencies": {
-    "@angular/compiler": "^9.1.0",
-    "@angular/core": "^9.1.0",
-    "document-register-element": "1.13.1",
-    "react": "16.12.0",
-    "react-dom": "16.12.0"
+    "react": "17.0.1",
+    "react-dom": "17.0.1",
+    "styled-jsx": "3.3.1",
+    "immer": "^8.0.1",
+    "@datorama/akita": "^6.1.2",
+    "rxjs": "6.5.5"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "~0.901.0",
-    "@angular-devkit/build-ng-packagr": "~0.901.0",
-    "@angular-devkit/core": "^9.1.0",
-    "@angular/compiler": "^9.1.0",
-    "@angular/compiler-cli": "^9.1.0",
-    "@angular/core": "^9.1.0",
-    "@babel/core": "7.8.4",
-    "@babel/preset-env": "7.8.4",
-    "@babel/preset-react": "7.8.3",
-    "@babel/preset-typescript": "7.8.3",
-    "@nrwl/angular": "^9.2.1",
-    "@nrwl/cypress": "9.2.1",
-    "@nrwl/eslint-plugin-nx": "9.2.1",
-    "@nrwl/express": "^9.2.1",
-    "@nrwl/jest": "9.2.1",
-    "@nrwl/nest": "^9.2.1",
-    "@nrwl/node": "^9.2.1",
-    "@nrwl/react": "9.2.1",
-    "@nrwl/tao": "^9.2.1",
-    "@nrwl/web": "^9.2.1",
-    "@nrwl/workspace": "9.2.1",
-    "@testing-library/react": "^10.0.2",
-    "@types/jest": "25.1.4",
-    "@types/node": "~8.9.4",
-    "@types/react": "16.9.17",
-    "@types/react-dom": "16.9.4",
-    "@typescript-eslint/eslint-plugin": "2.19.2",
-    "@typescript-eslint/parser": "2.19.2",
-    "babel-jest": "25.1.0",
-    "cypress": "^4.1.0",
+    "tslib": "^2.0.0",
+    "core-js": "^3.6.5",
+    "@ungap/custom-elements": "^0.1.15",
+    "@angular-devkit/build-angular": "0.1102.4",
+    "@angular-devkit/core": "10.1.7",
+    "@angular/compiler": "11.2.5",
+    "@angular/compiler-cli": "11.2.5",
+    "@angular/core": "11.2.5",
+    "@babel/core": "7.9.6",
+    "@babel/preset-env": "7.9.6",
+    "@babel/preset-react": "7.9.4",
+    "@babel/preset-typescript": "7.9.0",
+    "@ngrx/component-store": "11.0.0",
+    "@nrwl/angular": "11.5.1",
+    "@nrwl/cli": "11.5.1",
+    "@nrwl/cypress": "11.5.1",
+    "@nrwl/eslint-plugin-nx": "11.5.1",
+    "@nrwl/jest": "11.5.1",
+    "@nrwl/linter": "11.5.1",
+    "@nrwl/react": "11.5.1",
+    "@nrwl/tao": "11.5.1",
+    "@nrwl/web": "11.5.1",
+    "@nrwl/workspace": "11.5.1",
+    "@testing-library/react": "^11.2.5",
+    "@testing-library/react-hooks": "^5.1.0",
+    "@types/jest": "26.0.8",
+    "@types/node": "12.12.38",
+    "@types/react": "16.9.56",
+    "@types/react-dom": "16.9.9",
+    "@types/styled-jsx": "2.2.8",
+    "@typescript-eslint/eslint-plugin": "4.3.0",
+    "@typescript-eslint/parser": "4.3.0",
+    "babel-jest": "26.2.2",
+    "cypress": "^6.0.1",
     "dotenv": "6.2.0",
-    "eslint": "6.8.0",
-    "eslint-config-prettier": "6.0.0",
-    "eslint-plugin-import": "2.19.1",
-    "eslint-plugin-jsx-a11y": "6.2.3",
-    "eslint-plugin-react": "7.17.0",
-    "eslint-plugin-react-hooks": "2.3.0",
-    "jest": "25.2.3",
-    "jest-preset-angular": "8.1.2",
-    "ng-packagr": "^9.0.0",
-    "prettier": "1.19.1",
+    "eslint": "7.10.0",
+    "eslint-config-prettier": "8.1.0",
+    "eslint-plugin-cypress": "^2.10.3",
+    "eslint-plugin-import": "2.22.1",
+    "eslint-plugin-jsx-a11y": "6.4.1",
+    "eslint-plugin-react": "7.21.5",
+    "eslint-plugin-react-hooks": "4.2.0",
+    "jest": "26.2.2",
+    "ng-packagr": "11.2.4",
+    "prettier": "2.2.1",
     "rollup": "^2.3.2",
-    "ts-jest": "25.2.1",
-    "ts-node": "~7.0.0",
-    "tslint": "~6.0.0",
-    "typescript": "~3.8.3"
+    "ts-jest": "26.4.0",
+    "ts-node": "9.1.1",
+    "typescript": "~4.0.3"
   }
 }
