@@ -1,6 +1,35 @@
-import { Observable, Subject } from 'rxjs';
-import { first } from 'rxjs/operators';
+import { Observable } from 'rxjs';
+import { distinctUntilChanged, first, toArray } from 'rxjs/operators';
 
+/**
+ * @whatItDoes reads all the values from an observable and returns a promise
+ * with an array of all values. This should be used in combination with async/await.
+ * NOTE:  !!This requires the stream to 'complete'
+ *
+ * ## Example
+ *
+ * ```typescript
+ * const obs = of(1, 2, 3, 4);
+ * const res = await readAll(obs)
+ * expect(res).toEqual([1, 2, 3, 4]);
+ * ```
+ */
+export function readAll<T>(o: Observable<T>): Promise<T[]> {
+  return o.pipe(toArray()).toPromise();
+}
+
+/**
+ * @whatItDoes reads the first value from an observable and returns a promise
+ * with it. This should be used in combination with async/await.
+ *
+ * ## Example
+ *
+ * ```typescript
+ * const obs = of(1, 2, 3, 4);
+ * const res = await readFirst(obs)
+ * expect(res).toEqual(1);
+ * ```
+ */
 export function readFirst<T>(o: Observable<T>): Promise<T> {
   return o.pipe(first()).toPromise();
 }
